Refactor code, adjust increaseQuantity and decreaseQuantity functions
